Move definitions into a .cpp file
