+++ Makefile
@@ -1,12 +1,13 @@
 HEADER=simcpp.h protothread.h
+SOURCE=simcpp.cpp
 EXE=example-minimal example-twocars
 
 .PHONY: clean
 
 all: $(EXE)
 
-%: %.cpp $(HEADER)
-	g++ -Wall -std=c++11 $< -o $@
+%: %.cpp $(HEADER) $(SOURCE)
+	g++ -Wall -std=c++11 $< $(SOURCE) -o $@
 
 clean:
 	rm $(EXE)
